feat: improve session handling
Apparently the sessions are automatically refreshed on usage.
As we can't be certain about how much they will be extended for,
I added an additional call in batch to every call, that uses
session.list to get the currently left seconds.
Without this, the integration creates new sessions because it thinks the
old session is expired, which it in did not, which can result in "lots" of
open session on the openwrt device.