Improve device deletion logic
When the STA sensor is enabled only a limited number of interface
sensors will not be removed (eth0, br-lan, lan, wan). Instead of
continuing to have more and more edge cases based on the name, this
creates a new parent device for:
* Interfaces
* Access Points
* MWAN3

Then the deletion logic when these options are disabled is updated to
delete children of these new devices and this device. This means the STA
deletion logic only needs to exclude these top level "parent" devices.